// frontend/src/app/models/fullStats.ts
export interface FullStats {
  Season: string;
  Number_Of_Matches: number;
  Number_Red_Cards: number;
  Number_Yellow_Cards: number;
  Number_Play_Minutes: number;
  Passes_Accuracy: number;
  Number_Licenses: string;
  Saves_Match: number;
  Missed_Goal_Match: number;
  Percentage_Served_Penalty: number;
  Number_Interceptions_Match: number;
  Goals: number;
  Assists: number;
  Key_Pass: number;
  Hit_Match: number;
  Goal_Match: number;
  Hit_Goal: number;
  Number_Selections_Match: number;
  Number_Takeaways_Match: number;
}